+++ app/src/main/java/f/cking/software/ui/filter/FilterScreen.kt
@@ -3,11 +3,31 @@ package f.cking.software.ui.filter
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.*
+import androidx.compose.material.Checkbox
+import androidx.compose.material.Chip
+import androidx.compose.material.ChipDefaults
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.material.TextField
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Clear
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.List
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -22,20 +42,22 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.vanpra.composematerialdialogs.rememberMaterialDialogState
-import f.cking.software.*
 import f.cking.software.R
-import f.cking.software.common.ClickableField
-import f.cking.software.common.RoundedBox
-import f.cking.software.common.TagChip
-import f.cking.software.common.navigation.Router
-import f.cking.software.common.rememberDateDialog
-import f.cking.software.common.rememberTimeDialog
+import f.cking.software.dateTimeFormat
+import f.cking.software.dateTimeStringFormat
+import f.cking.software.toLocalTime
+import f.cking.software.toMilliseconds
 import f.cking.software.ui.ScreenNavigationCommands
 import f.cking.software.ui.tagdialog.TagDialog
+import f.cking.software.utils.graphic.ClickableField
+import f.cking.software.utils.graphic.RoundedBox
+import f.cking.software.utils.graphic.TagChip
+import f.cking.software.utils.graphic.rememberDateDialog
+import f.cking.software.utils.graphic.rememberTimeDialog
+import f.cking.software.utils.navigation.Router
 import java.time.LocalDate
 import java.time.LocalTime
 import java.time.ZoneId
-import java.util.*
 import java.util.concurrent.TimeUnit
 
 object FilterScreen {
