+++ app/src/main/java/f/cking/software/ui/journal/JournalScreen.kt
@@ -29,9 +29,9 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.google.accompanist.flowlayout.FlowRow
 import f.cking.software.R
-import f.cking.software.common.BottomSpacer
-import f.cking.software.common.ContentPlaceholder
-import f.cking.software.common.Divider
+import f.cking.software.utils.graphic.BottomSpacer
+import f.cking.software.utils.graphic.ContentPlaceholder
+import f.cking.software.utils.graphic.Divider
 import org.koin.androidx.compose.koinViewModel
 
 object JournalScreen {
