+++ app/src/main/java/f/cking/software/ui/journal/JournalViewModel.kt
@@ -8,13 +8,13 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import f.cking.software.R
-import f.cking.software.common.navigation.Router
 import f.cking.software.data.repo.DevicesRepository
 import f.cking.software.data.repo.JournalRepository
 import f.cking.software.data.repo.RadarProfilesRepository
 import f.cking.software.dateTimeStringFormat
 import f.cking.software.domain.model.JournalEntry
 import f.cking.software.ui.ScreenNavigationCommands
+import f.cking.software.utils.navigation.Router
 import kotlinx.coroutines.launch
 
 class JournalViewModel(
