+++ app/src/main/java/f/cking/software/ui/main/MainScreen.kt
@@ -29,9 +29,9 @@ import com.vanpra.composematerialdialogs.MaterialDialog
 import com.vanpra.composematerialdialogs.MaterialDialogState
 import com.vanpra.composematerialdialogs.rememberMaterialDialogState
 import f.cking.software.R
-import f.cking.software.common.BlurredNavBar
-import f.cking.software.common.pxToDp
 import f.cking.software.ui.GlobalUiState
+import f.cking.software.utils.graphic.BlurredNavBar
+import f.cking.software.utils.graphic.pxToDp
 import org.koin.androidx.compose.koinViewModel
 
 object MainScreen {
