+++ app/src/main/java/f/cking/software/ui/profiledetails/ProfileDetailsViewModel.kt
@@ -8,14 +8,14 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import f.cking.software.R
-import f.cking.software.common.navigation.BackCommand
-import f.cking.software.common.navigation.Router
 import f.cking.software.data.repo.RadarProfilesRepository
 import f.cking.software.domain.interactor.DeleteRadarProfile
 import f.cking.software.domain.interactor.SaveRadarProfile
 import f.cking.software.domain.model.RadarProfile
 import f.cking.software.ui.filter.FilterUiMapper
 import f.cking.software.ui.filter.FilterUiState
+import f.cking.software.utils.navigation.BackCommand
+import f.cking.software.utils.navigation.Router
 import kotlinx.coroutines.launch
 
 class ProfileDetailsViewModel(
