+++ app/src/main/java/f/cking/software/utils/navigation/NavigationCommand.kt
@@ -1,4 +1,4 @@
-package f.cking.software.common.navigation
+package f.cking.software.utils.navigation
 
 import androidx.compose.runtime.Composable
 import java.util.UUID
