+++ app/src/main/java/f/cking/software/utils/navigation/Router.kt
@@ -1,4 +1,4 @@
-package f.cking.software.common.navigation
+package f.cking.software.utils.navigation
 
 interface Router {
     fun navigate(command: NavigationCommand)
